move arklet logger to koupleless common
