+++ arklet-core/src/main/java/com/alipay/sofa/koupleless/arklet/core/ArkletComponentRegistry.java
@@ -25,8 +25,7 @@ import com.alipay.sofa.koupleless.arklet.core.health.HealthServiceImpl;
 import com.alipay.sofa.koupleless.arklet.core.api.ApiClient;
 import com.alipay.sofa.koupleless.arklet.core.command.CommandService;
 import com.alipay.sofa.koupleless.arklet.core.command.CommandServiceImpl;
-import com.alipay.sofa.koupleless.arklet.core.common.log.ArkletLogger;
-import com.alipay.sofa.koupleless.arklet.core.common.log.ArkletLoggerFactory;
+import com.alipay.sofa.koupleless.common.log.ArkletLoggerFactory;
 import com.alipay.sofa.koupleless.arklet.core.ops.UnifiedOperationService;
 import com.alipay.sofa.koupleless.arklet.core.ops.UnifiedOperationServiceImpl;
 import com.google.inject.AbstractModule;
