+++ arklet-core/src/main/java/com/alipay/sofa/koupleless/arklet/core/command/CommandServiceImpl.java
@@ -45,8 +45,8 @@ import com.alipay.sofa.koupleless.arklet.core.command.record.ProcessRecord;
 import com.alipay.sofa.koupleless.arklet.core.command.record.ProcessRecordHolder;
 import com.alipay.sofa.koupleless.arklet.core.common.exception.ArkletInitException;
 import com.alipay.sofa.koupleless.arklet.core.common.exception.CommandValidationException;
-import com.alipay.sofa.koupleless.arklet.core.common.log.ArkletLogger;
-import com.alipay.sofa.koupleless.arklet.core.common.log.ArkletLoggerFactory;
+import com.alipay.sofa.koupleless.common.log.ArkletLogger;
+import com.alipay.sofa.koupleless.common.log.ArkletLoggerFactory;
 import com.alipay.sofa.koupleless.arklet.core.util.AssertUtils;
 import com.google.inject.Singleton;
