--- koupleless-common/src/main/resources/com/alipay/sofa/koupleless/log/log4j/log-conf.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
-
-<log4j:configuration xmlns:log4j="http://jakarta.apache.org/log4j/">
-    <appender name="ERROR-APPENDER" class="org.apache.log4j.DailyRollingFileAppender">
-        <param name="file" value="${logging.path}/sofa-koupleless/common-error.log"/>
-        <param name="append" value="true"/>
-        <param name="encoding" value="${file.encoding}"/>
-        <param name="threshold" value="error"/>
-        <layout class="org.apache.log4j.PatternLayout">
-            <param name="ConversionPattern" value="%d %-5p %-32t - %m%n"/>
-        </layout>
-    </appender>
-
-    <appender name="SOFA-APPENDER" class="org.apache.log4j.DailyRollingFileAppender">
-        <param name="file" value="${logging.path}/sofa-koupleless/common-default.log"/>
-        <param name="append" value="true"/>
-        <param name="encoding" value="${file.encoding}"/>
-        <param name="threshold" value="${logging.level.com.alipay.sofa}"/>
-        <layout class="org.apache.log4j.PatternLayout">
-            <param name="ConversionPattern" value="%d %-5p %-32t - %m%n"/>
-        </layout>
-    </appender>
-
-    <appender name="CONSOLE-APPENDER" class="org.apache.log4j.ConsoleAppender">
-        <layout class="org.apache.log4j.PatternLayout">
-            <param name="ConversionPattern" value="%d %-5p %-32t - %m%n"/>
-        </layout>
-    </appender>
-
-    <logger name="com.alipay.sofa.koupleless" additivity="false">
-        <level value="${logging.level.com.alipay.sofa}"/>
-        <appender-ref ref="SOFA-APPENDER"/>
-        <appender-ref ref="ERROR-APPENDER"/>
-        <appender-ref ref="CONSOLE-APPENDER"/>
-    </logger>
-
-    <root>
-        <level value="${KOUPLELESS_LOG_LEVEL}"/>
-        <appender-ref ref="SOFA-APPENDER"/>
-        <appender-ref ref="ERROR-APPENDER"/>
-        <appender-ref ref="CONSOLE-APPENDER"/>
-    </root>
-</log4j:configuration>
