--- koupleless-common/src/main/resources/com/alipay/sofa/koupleless/log/log4j2/log-conf.xml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration status="OFF">
-    <Properties>
-        <Property name="KOUPLELESS_LOG_LEVEL">${sys:logging.level.com.alipay.sofa}</Property>
-        <Property name="KOUPLELESS_LOG_PATH">${sys:logging.path}</Property>
-        <Property name="KOUPLELESS_FILE_ENCODING">${sys:file.encoding}</Property>
-    </Properties>
-
-    <Appenders>
-        <RollingFile name="ERROR-APPENDER" fileName="${KOUPLELESS_LOG_PATH}/sofa-koupleless/common-error.log" append="true"
-                     filePattern="${KOUPLELESS_LOG_PATH}/sofa-koupleless/common-error.log.%d{yyyy-MM-dd}">
-            <ThresholdFilter level="error" onMatch="ACCEPT" onMismatch="DENY"/>
-            <PatternLayout charset="${KOUPLELESS_FILE_ENCODING}">
-                <pattern>%d %-5p %-32t - %m%n</pattern>
-            </PatternLayout>
-            <Policies>
-                <TimeBasedTriggeringPolicy interval="1" modulate="true"/>
-            </Policies>
-        </RollingFile>
-
-        <RollingFile name="SOFA-APPENDER" fileName="${KOUPLELESS_LOG_PATH}/sofa-koupleless/common-default.log" append="true"
-                     filePattern="${KOUPLELESS_LOG_PATH}/sofa-koupleless/common-default.log.%d{yyyy-MM-dd}">
-            <ThresholdFilter level="${KOUPLELESS_LOG_LEVEL}" onMatch="ACCEPT" onMismatch="DENY"/>
-            <PatternLayout charset="${KOUPLELESS_FILE_ENCODING}">
-                <pattern>%d %-5p %-32t - %m%n</pattern>
-            </PatternLayout>
-            <Policies>
-                <TimeBasedTriggeringPolicy interval="1" modulate="true"/>
-            </Policies>
-        </RollingFile>
-
-        <Console name="CONSOLE-APPENDER" target="SYSTEM_OUT">
-            <PatternLayout charset="${KOUPLELESS_FILE_ENCODING}">
-                <pattern>%d %-5p %-32t - %m%n</pattern>
-            </PatternLayout>
-        </Console>
-
-    </Appenders>
-
-    <Loggers>
-        <Logger name="com.alipay.sofa.koupleless" level="${KOUPLELESS_LOG_LEVEL}" additivity="false">
-            <appender-ref ref="SOFA-APPENDER"/>
-            <appender-ref ref="ERROR-APPENDER"/>
-            <appender-ref ref="CONSOLE-APPENDER"/>
-        </Logger>
-
-        <Root level="${KOUPLELESS_LOG_LEVEL}">
-            <appender-ref ref="SOFA-APPENDER"/>
-            <appender-ref ref="ERROR-APPENDER"/>
-            <appender-ref ref="CONSOLE-APPENDER"/>
-        </Root>
-    </Loggers>
-</configuration>
