--- koupleless-common/src/main/resources/com/alipay/sofa/koupleless/log/logback/log-conf.xml
@@ -1,76 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Licensed to the Apache Software Foundation (ASF) under one or more
-  ~ contributor license agreements.  See the NOTICE file distributed with
-  ~ this work for additional information regarding copyright ownership.
-  ~ The ASF licenses this file to You under the Apache License, Version 2.0
-  ~ (the "License"); you may not use this file except in compliance with
-  ~ the License.  You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<configuration>
-    <include resource="org/springframework/boot/logging/logback/defaults.xml" />
-    <property name="LOG_PATTERN" value="%clr(%d{yyyy-MM-dd HH:mm:ss.SSS}){faint} %clr(${LOG_LEVEL_PATTERN:-%5p}) %clr(${PID:- }){magenta} %clr(---){faint} %clr([%15.15t]){faint} %clr(com.alipay.sofa.koupleless){cyan} %clr(:){faint} %m%n${LOG_EXCEPTION_CONVERSION_WORD:-%wEx}" />
-    <appender name="ERROR-APPENDER" class="ch.qos.logback.core.rolling.RollingFileAppender">
-        <append>true</append>
-        <filter class="ch.qos.logback.classic.filter.ThresholdFilter">
-            <level>error</level>
-            <onMatch>ACCEPT</onMatch>
-            <onMismatch>DENY</onMismatch>
-        </filter>
-        <file>${logging.path}/sofa-koupleless/common-error.log</file>
-        <rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">
-            <FileNamePattern>${logging.path}/sofa-koupleless/common-error.log.%d{yyyy-MM-dd}</FileNamePattern>
-            <MaxHistory>30</MaxHistory>
-        </rollingPolicy>
-        <encoder class="ch.qos.logback.classic.encoder.PatternLayoutEncoder">
-            <pattern>%d %-5p %-32t - %m%n</pattern>
-            <charset>${file.encoding}</charset>
-        </encoder>
-    </appender>
-
-    <appender name="SOFA-APPENDER" class="ch.qos.logback.core.rolling.RollingFileAppender">
-        <append>true</append>
-        <filter class="ch.qos.logback.classic.filter.ThresholdFilter">
-            <level>${logging.level.com.alipay.sofa}</level>
-            <onMatch>ACCEPT</onMatch>
-            <onMismatch>DENY</onMismatch>
-        </filter>
-        <file>${logging.path}/sofa-koupleless/common-default.log</file>
-        <rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">
-            <FileNamePattern>${logging.path}/sofa-koupleless/common-default.log.%d{yyyy-MM-dd}</FileNamePattern>
-            <MaxHistory>30</MaxHistory>
-        </rollingPolicy>
-        <encoder class="ch.qos.logback.classic.encoder.PatternLayoutEncoder">
-            <pattern>%d %-5p %-32t - %m%n</pattern>
-            <charset>${file.encoding}</charset>
-        </encoder>
-    </appender>
-
-    <appender name="CONSOLE-APPENDER" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>${LOG_PATTERN}</pattern>
-            <charset>utf8</charset>
-        </encoder>
-    </appender>
-
-    <logger name="com.alipay.sofa.koupleless" level="${logging.level.com.alipay.sofa}" additivity="false">
-        <appender-ref ref="SOFA-APPENDER"/>
-        <appender-ref ref="ERROR-APPENDER"/>
-        <appender-ref ref="CONSOLE-APPENDER"/>
-    </logger>
-
-    <root level="${KOUPLELESS_LOG_LEVEL}">
-        <appender-ref ref="SOFA-APPENDER"/>
-        <appender-ref ref="ERROR-APPENDER"/>
-        <appender-ref ref="CONSOLE-APPENDER"/>
-    </root>
-</configuration>
