+++ koupleless-ext/koupleless-ext-autoconfigure/pom.xml
@@ -58,6 +58,7 @@
         <dependency>
             <groupId>org.mockito</groupId>
             <artifactId>mockito-inline</artifactId>
+            <version>4.5.1</version>
             <scope>test</scope>
         </dependency>
         <dependency>
